+++ XADMaster/CSHandle.h
@@ -48,8 +48,8 @@ extern NSExceptionName const CSNotSupportedException;
 -(void)seekToFileOffset:(off_t)offs;
 -(void)seekToEndOfFile;
 -(void)pushBackByte:(int)byte;
--(int)readAtMost:(int)num toBuffer:(void *)buffer;
--(void)writeBytes:(int)num fromBuffer:(const void *)buffer;
+-(NSInteger)readAtMost:(NSInteger)num toBuffer:(void *)buffer;
+-(void)writeBytes:(NSInteger)num fromBuffer:(const void *)buffer;
 
 
 
@@ -95,11 +95,11 @@ extern NSExceptionName const CSNotSupportedException;
 
 -(NSData *)fileContents;
 -(NSData *)remainingFileContents;
--(NSData *)readDataOfLength:(int)length;
--(NSData *)readDataOfLengthAtMost:(int)length;
--(NSData *)copyDataOfLength:(int)length;
--(NSData *)copyDataOfLengthAtMost:(int)length;
--(void)readBytes:(int)num toBuffer:(void *)buffer;
+-(NSData *)readDataOfLength:(NSInteger)length;
+-(NSData *)readDataOfLengthAtMost:(NSInteger)length;
+-(NSData *)copyDataOfLength:(NSInteger)length;
+-(NSData *)copyDataOfLengthAtMost:(NSInteger)length;
+-(void)readBytes:(NSInteger)num toBuffer:(void *)buffer;
 
 -(off_t)readAndDiscardAtMost:(off_t)num;
 -(void)readAndDiscardBytes:(off_t)num;
