+++ XADMaster/Checksums.h
@@ -38,8 +38,8 @@
 -(void)seekToFileOffset:(off_t)offs;
 -(void)seekToEndOfFile;
 -(void)pushBackByte:(int)byte;
--(int)readAtMost:(int)num toBuffer:(void *)buffer;
--(void)writeBytes:(int)num fromBuffer:(const void *)buffer;
+-(NSInteger)readAtMost:(NSInteger)num toBuffer:(void *)buffer;
+-(void)writeBytes:(NSInteger)num fromBuffer:(const void *)buffer;
 
 @property (NS_NONATOMIC_IOSONLY, readonly) BOOL hasChecksum;
 @property (NS_NONATOMIC_IOSONLY, readonly, getter=isChecksumCorrect) BOOL checksumCorrect;
