+++ XADMaster/XADBlockHandle.h
@@ -22,6 +22,6 @@ firstBlock:(uint32_t)first headerSize:(off_t)headersize;
 
 -(void)seekToFileOffset:(off_t)offs;
 -(void)seekToEndOfFile;
--(int)readAtMost:(int)num toBuffer:(void *)buffer;
+-(NSInteger)readAtMost:(NSInteger)num toBuffer:(void *)buffer;
 
 @end
